# Conflicts:
#	feature-route/impl/src/client/kotlin/orderchanges/presentation/compose/view/ConfirmOrderChangesView.kt
#	feature-route/impl/src/main/kotlin/root/data/model/response/OrderExtrasDto.kt
#	feature-route/impl/src/main/kotlin/root/domain/model/extras/OrderExtrasModel.kt

+++ feature-route/impl/src/contractor/kotlin/root/domain/model/RouteModel.kt
@@ -1,9 +1,11 @@
 package root.domain.model
 
+import java.math.BigDecimal
+
 data class RouteModel(
     val id: Long,
     val orders: List<RouteOrderModel>,
-    val price: Int,
+    val price: BigDecimal,
     val distance: Double,
     val status: RouteStatusProgress
 )
